requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
google-generativeai
openai